PRESIDECMS-3017 Update to render and handle form page submit.

+++ system/services/formbuilder/FormBuilderService.cfc
@@ -76,10 +76,18 @@ component {
 	 * @autodoc    true
 	 * @id.hint    ID of the form item's form you wish to get
 	 */
-	public array function getFormItemDefaultFields( required string id ) {
+	public array function getFormItemDefaultFields(
+		  required string  id
+		,          boolean withPageNumber = false
+	) {
 		var fields = [ "id", "item_type", "configuration", "form" ];
+
 		if ( isV2Form( formid=arguments.id ) ) {
 			ArrayAppend( fields, "question" );
+		}
+
+		if ( arguments.withPageNumber ) {
+			ArrayAppend( fields, "count( case when item_type = 'page' then 1 end) over ( order by sort_order rows between unbounded preceding and current row ) as page_number" );
 		}
 
 		return fields;
@@ -92,16 +100,20 @@ component {
 	 * @id.hint        ID of the form whose sections and items you wish to get
 	 * @itemTypes.hint Optional array of item types with which to filter the returned form items
 	 */
-	public array function getFormItems( required string id, array itemTypes=[] ) {
+	public array function getFormItems(
+		  required string  id
+		,          array   itemTypes  = []
+		,          numeric pageNumber = 0
+	) {
 		var result = [];
 		var items  = $getPresideObject( "formbuilder_formitem" ).selectData(
-			  filter       = { form=arguments.id }
+			  selectFields = getFormItemDefaultFields( id=arguments.id, withPageNumber=$helpers.isTrue( arguments.pageNumber ) )
+			, filter       = { form=arguments.id }
 			, orderBy      = "sort_order"
-			, selectFields = getFormItemDefaultFields( id=arguments.id )
 		);
 
-		for( var item in items ) {
-			if ( !itemTypes.len() || itemTypes.findNoCase( item.item_type ) ) {
+		for ( var item in items ) {
+			if ( ( arguments.pageNumber == 0 || arguments.pageNumber == item.page_number ) && ( !ArrayLen( itemTypes ) || ArrayFindNoCase( itemTypes, item.item_type ) ) ) {
 				var preparedItem = {
 					  id            = item.id
 					, formId        = item.form
@@ -687,13 +699,13 @@ component {
 	 * @formId.hint     The ID of the form you wish to retrieve stored values for
 	 *
 	 */
-	public struct function getTempStoredSubmission( required string formId ) {
+	public struct function getTempStoredSubmission( required string formId, boolean clearSubmission=false ) {
 		var tempStorageKey = "temp_formbuilder_submission_#formId#";
 		var submission     =  _getSessionStorage().getVar( tempStorageKey, StructNew() );
 
-
-		_getSessionStorage().deleteVar( tempStorageKey );
-
+		if ( arguments.clearSubmission ) {
+			_getSessionStorage().deleteVar( tempStorageKey );
+		}
 
 		return submission;
 	}
@@ -714,8 +726,12 @@ component {
 		,          struct configuration    = {}
 		,          any    validationResult = ""
 	) {
+		var formPersistData = $getRequestContext().getCollectionWithoutSystemVars();
+		arguments.configuration.formPagesCount = countFormPages( formId=arguments.formId );
+		arguments.configuration.formPageNumber = arguments.configuration.formPagesCount ? ( formPersistData.formPageNumber ?: 1 ) : 0;
+
 		var formConfiguration = getForm( id=arguments.formId );
-		var items             = getFormItems( id=arguments.formId );
+		var items             = getFormItems( id=arguments.formId, pageNumber=arguments.configuration.formPageNumber );
 		var renderedItems     = CreateObject( "java", "java.lang.StringBuffer" );
 		var coreLayoutArgs    = Duplicate( arguments.configuration );
 		var coreLayoutViewlet = "formbuilder.core.formLayout";
@@ -723,9 +739,9 @@ component {
 		var formLayoutViewlet = _getFormBuilderRenderingService().getFormLayoutViewlet( layout=arguments.layout );
 		var idPrefixForFields = _createIdPrefix( formId=arguments.formId );
 
-		for( var item in items ) {
+		for ( var item in items ) {
 			var config    = Duplicate( item.configuration );
-			var fieldName = config.name ?: CreateUUId();
+			var fieldName = config.name ?: CreateUUID();
 
 			config.id = idPrefixForFields & fieldName;
 
@@ -745,15 +761,15 @@ component {
 		coreLayoutArgs.renderedItems = renderedItems.toString();
 		coreLayoutArgs.id            = idPrefixForFields;
 		coreLayoutArgs.formItems     = items;
+
 		for( var f in formConfiguration ) {
 			coreLayoutArgs.configuration = f;
 		}
 
-		formLayoutArgs.renderedForm  = $renderViewlet(
+		formLayoutArgs.renderedForm = $renderViewlet(
 			  event = coreLayoutViewlet
 			, args  = coreLayoutArgs
 		);
-
 
 		return $renderViewlet(
 			  event = formLayoutViewlet
@@ -2160,6 +2176,13 @@ component {
 	public boolean function isV2Form( required string formid ) {
 		return $isFeatureEnabled( "formbuilder2" ) && $getPresideObject( "formbuilder_form" ).dataExists(
 			  filter = { id=arguments.formId, uses_global_questions=true }
+		);
+	}
+
+	public numeric function countFormPages( required string formId ) {
+		return $getPresideObject( "formbuilder_formitem" ).selectData(
+			  filter          = { form=arguments.formId, item_type="page" }
+			, recordCountOnly = true
 		);
 	}
